+++ ME3Explorer/ME3Explorer.csproj
@@ -666,13 +666,14 @@
     </Compile>
     <Compile Include="SharedUI\ListScrollBox.cs" />
     <Compile Include="SharedUI\ObservableCollectionExtended.cs" />
+    <Compile Include="SharedUI\PeregrineTreeView\TreeViewHelper.cs" />
     <Compile Include="SharedUI\PromptDialog.xaml.cs">
       <DependentUpon>PromptDialog.xaml</DependentUpon>
     </Compile>
     <Compile Include="SharedUI\StretchingTreeView.cs" />
     <Compile Include="SharedUI\PeregrineTreeView\DispatcherHelper.cs" />
     <Compile Include="SharedUI\PeregrineTreeView\PeregrineHeap.cs" />
-    <Compile Include="SharedUI\PeregrineTreeView\TreeViewHelper.cs" />
+    <Compile Include="SharedUI\PeregrineTreeView\TreeViewItemHelper.cs" />
     <Compile Include="SingleInstance.cs" />
     <Compile Include="Soundplorer\ISACT_Parser.cs" />
     <Compile Include="Soundplorer\ISBank.cs" />
