+++ ME3Explorer/PackageEditor/PackageEditorWPF.xaml
@@ -288,8 +288,8 @@
                         <Style TargetType="{x:Type TreeViewItem}">
                             <Setter Property="IsSelected" Value="{Binding Path=IsSelected, Mode=TwoWay}" />
                             <Setter Property="IsExpanded" Value="{Binding Path=IsExpanded, Mode=TwoWay}" />
-                            <Setter Property="vhelp:TreeViewHelper.BringSelectedItemIntoView" Value="True" />
-                            <Setter Property="vhelp:TreeViewHelper.BringExpandedChildrenIntoView" Value="True" />
+                            <Setter Property="vhelp:TreeViewItemHelper.BringSelectedItemIntoView" Value="True" />
+                            <Setter Property="vhelp:TreeViewItemHelper.BringExpandedChildrenIntoView" Value="True" />
                         </Style>
                     </TreeView.ItemContainerStyle>
                     <TreeView.ContextMenu>
@@ -306,6 +306,9 @@
                             </MenuItem>
                         </ContextMenu>
                     </TreeView.ContextMenu>
+                    <i:Interaction.Behaviors>
+                        <vhelp:TreeViewHelper BoundSelectedItem="{Binding SelectedItem}" />
+                    </i:Interaction.Behaviors>
                 </TreeView>
                 <GridSplitter Grid.Column="1" Width="5" HorizontalAlignment="Stretch" Grid.Row="1"/>
                 <TabControl Name="EditorTabs" Grid.Column="2" Grid.Row="1">
