+++ ME3Explorer/SharedUI/PeregrineTreeView/TreeViewHelper.cs
@@ -1,120 +1,46 @@
-﻿using System;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Threading;
+using System.Windows.Interactivity;
 
 namespace ME3Explorer.SharedUI.PeregrineTreeView
 {
-    public static class TreeViewHelper
+    public class TreeViewHelper : Behavior<TreeView>
     {
-        public static bool GetBringSelectedItemIntoView(TreeViewItem treeViewItem)
+        public object BoundSelectedItem
         {
-            return (bool)treeViewItem.GetValue(BringSelectedItemIntoViewProperty);
+            get => GetValue(BoundSelectedItemProperty);
+            set => SetValue(BoundSelectedItemProperty, value);
         }
 
-        public static void SetBringSelectedItemIntoView(TreeViewItem treeViewItem, bool value)
-        {
-            treeViewItem.SetValue(BringSelectedItemIntoViewProperty, value);
-        }
-
-        public static readonly DependencyProperty BringSelectedItemIntoViewProperty =
-            DependencyProperty.RegisterAttached(
-                "BringSelectedItemIntoView",
-                typeof(bool),
+        public static readonly DependencyProperty BoundSelectedItemProperty =
+            DependencyProperty.Register("BoundSelectedItem",
+                typeof(object),
                 typeof(TreeViewHelper),
-                new UIPropertyMetadata(false, BringSelectedItemIntoViewChanged));
+                new FrameworkPropertyMetadata(null,
+                    FrameworkPropertyMetadataOptions.BindsTwoWayByDefault,
+                    OnBoundSelectedItemChanged));
 
-        private static void BringSelectedItemIntoViewChanged(DependencyObject obj, DependencyPropertyChangedEventArgs args)
+        private static void OnBoundSelectedItemChanged(DependencyObject obj, DependencyPropertyChangedEventArgs args)
         {
-            if (!(args.NewValue is bool))
-                return;
-
-            if (!(obj is TreeViewItem item))
-                return;
-
-            if ((bool)args.NewValue)
-                item.Selected += OnTreeViewItemSelected;
-            else
-                item.Selected -= OnTreeViewItemSelected;
+            if (args.NewValue is TreeViewEntry item)
+                item.IsSelected = true;
         }
 
-        private static void OnTreeViewItemSelected(object sender, RoutedEventArgs e)
+        protected override void OnAttached()
         {
-            // prevent this event bubbling up to any parent nodes
-            e.Handled = true;
-
-            if (sender is TreeViewItem item)
-            {
-                // use DispatcherPriority.ApplicationIdle so this occurs after all operations related to tree item expansion
-                DispatcherHelper.AddToQueue(() =>
-                {
-                    System.Diagnostics.Debug.WriteLine("BRINGING INTO VIEW - SELECTED NODE!");
-                    item.BringIntoView();
-                }, DispatcherPriority.ApplicationIdle);
-            }
+            base.OnAttached();
+            AssociatedObject.SelectedItemChanged += OnTreeViewSelectedItemChanged;
         }
 
-        public static bool GetBringExpandedChildrenIntoView(TreeViewItem treeViewItem)
+        protected override void OnDetaching()
         {
-            return (bool)treeViewItem.GetValue(BringExpandedChildrenIntoViewProperty);
+            AssociatedObject.SelectedItemChanged -= OnTreeViewSelectedItemChanged;
+            base.OnDetaching();
         }
 
-        public static void SetBringExpandedChildrenIntoView(TreeViewItem treeViewItem, bool value)
+        private void OnTreeViewSelectedItemChanged(object obj, RoutedPropertyChangedEventArgs<object> args)
         {
-            treeViewItem.SetValue(BringExpandedChildrenIntoViewProperty, value);
-        }
-
-        public static readonly DependencyProperty BringExpandedChildrenIntoViewProperty =
-            DependencyProperty.RegisterAttached(
-                "BringExpandedChildrenIntoView",
-                typeof(bool),
-                typeof(TreeViewHelper),
-                new UIPropertyMetadata(false, BringExpandedChildrenIntoViewChanged));
-
-        private static void BringExpandedChildrenIntoViewChanged(DependencyObject obj, DependencyPropertyChangedEventArgs args)
-        {
-            if (!(args.NewValue is bool))
-                return;
-
-            if (!(obj is TreeViewItem item))
-                return;
-
-            if ((bool)args.NewValue)
-            {
-                item.Expanded += OnTreeViewItemExpanded;
-            }
-            else
-                item.Expanded -= OnTreeViewItemExpanded;
-        }
-
-        private static void OnTreeViewItemExpanded(object sender, RoutedEventArgs e)
-        {
-            // prevent this event bubbling up to any parent nodes
-            e.Handled = true;
-
-            if (!(sender is TreeViewItem item))
-                return;
-
-            // use DispatcherPriority.ContextIdle for all actions related to tree item expansion
-            // this ensures that all UI elements for any newly visible children are created before any selection operation
-
-            // first bring the last child into view
-            Action action = () =>
-            {
-                var lastChild = item.ItemContainerGenerator.ContainerFromIndex(item.Items.Count - 1) as TreeViewItem;
-                if (lastChild == null) { System.Diagnostics.Debug.WriteLine("FAIL!"); }
-                lastChild?.BringIntoView();
-            };
-
-            DispatcherHelper.AddToQueue(action, DispatcherPriority.ContextIdle);
-
-            // then bring the expanded item (back) into view
-            action = () => {
-                System.Diagnostics.Debug.WriteLine("BRINGING INTO VIEW - EXPANDED NODE FIRST: " + item.DataContext);
-                item.BringIntoView();
-            };
-
-            DispatcherHelper.AddToQueue(action, DispatcherPriority.ContextIdle);
+            BoundSelectedItem = args.NewValue;
         }
     }
 }
